refactor(settings+types): :truck: Move `settings.rs` to `types.rs`
This is in preparation of the new public API
